Updated NLU error  codes to 500 - internal server error. Updated middleware to have only parse api
