+++ package.json
@@ -22,13 +22,13 @@
     "gulp": "^3.9.1"
   },
   "config": {
-    "rasanluendpoint": "http://localhost:5000",
+    "rasanluendpoint": "http://ai-nlu-is360-chatbot-nonprod.ose-ctc-core.optum.com",
     "rasacoreendpoint": "http://localhost:5005",
     "coresecuritytoken": "",
     "nlusecuritytoken": "",
     "cacheagents": false,
     "jwtsecret": "th1s1sn0t0nlyths3cr3tbutth3m0st1mp0rtAntp13c30f1nf0rmAt10nf0rth1sAppl1cAt10n",
-    "postgresConnectionString": "postgres://postgres:rasaui@10.211.55.3:5432/rasaui"
+    "postgresConnectionString": "postgres://postgres:rasaui@apsrd8588:5432/rasaui"
   },
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
